backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
PyMySQL==1.1.0
cryptography==41.0.7

firebase-admin==6.3.0
python-dotenv==1.0.0

marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2

APScheduler==3.10.4
Flask-Limiter==3.5.0
redis==5.0.1

celery==5.3.4

python-jose[cryptography]==3.3.0
passlib==1.7.4

gunicorn==21.2.0